1.add InitialPage,FinishReinitializationPage,SelectDiskModePage and SetPasswordPage 2.finish implement ReinitializationPresenter 3.modify reinitialization ui effect according to the effect diagram

+++ app/src/main/java/com/winsun/fruitmix/newdesign201804/equipment/abnormal/HandleAbnormalEquipmentPresenter.kt
@@ -7,7 +7,7 @@ import com.winsun.fruitmix.R
 import com.winsun.fruitmix.recyclerview.BaseRecyclerViewAdapter
 import com.winsun.fruitmix.recyclerview.SimpleViewHolder
 
-data class DiskItem(val brand: String, val capacity: Long)
+data class DiskItem(val brand: String, val capacity: Double)
 
 class HandleAbnormalEquipmentPresenter {
